Remove detailed compiler check
Currently the Makefile tries to pick the latest version of g++ and
overrides CXX when found.

Since we give an error if the version of the compiler is not sufficient,
we can remove this check.

It's now possible to specify an older version (on purpose) and new
compiler version don't need to be added manually.