Set FreeBSD workflow timeout
Recently the FreeBSD workflow has started to hang in a boot loop when
the VM starts up. The issue is being tracked upstream but there is not
response at the moment.

To work around this set a timeout to not waste CI minutes. Other
workflows might also want this change since they don't take 20 minutes
anyway.