Install README in its own subdirectory
According to the Linux Standard Base (LSB), README files should be
installed in the /usr/share/doc directory. By convention however the
README files should not go there 'bare', but in a subdirectory for each
program.

Install to $(PREFIX)/share/doc/btop/README.md to comply with this.

To ensure the 'themes' is copied as a full directory and not individual
files, create the /usr/share/btop in a separate step as the doc
installation step no longer does it.