Fix hide and info buttons
This is partially my fault from when I made a change to allow
the hide button to not be greyed out when the detailed view item
was selected in the list.

I didn't realize that this deactivated the info button on the bottom bar from operating.

to fix this I have adding an additinal mouse mapping called "info_enter".
If there is another way to have to mouse mappings that is preferrable to this
I can do that instead.

Additionally I didn't realize that to disable a mouse mapping you have to 'delete' it so
the hide button was continuing to act as a button even when greyed out if at any point the
detailed view process was selected while detailed view was open.