Reassign tree-children-toggle keybind
the tree-children-toggle feature introduced in f8b4217 from PR #1274
was using the "u" key as a binding.

This conflicts with pause using the "u" key now.

Since the "u" key was chosen without any real reason for the
tree-children-toggle feature. It will be changed to "C" and
"u" can safely be used for pause.